# Handle nominator smaller than divisor in udivrem()