@holiman 2022-03-15
It has nothing to do with the argument aliasiing. Here's a repro: 
```
func TestBar(t *testing.T) {
	argA := new(Int)
	argB := new(Int)
	argA.fromHex("0x66aa0a5319bcf5cb4")
	argB.fromHex("0x1ce97e1ab91a")
	z := new(Int)
	modraw(z, argB.Clone(), argA.Clone())
	fmt.Printf("z: %x\n", z)

	b1a, b2a := argA.ToBig(), argB.ToBig()

	bigintFunc := func(b1, b2, b3 *big.Int) {
		if b3.Sign() == 0 {
			b1.SetUint64(0)
		} else {
			b1.Mod(b2, b3)
		}
	}

	bigintFunc(b1a, b2a, b1a)
	fmt.Printf("z: %x\n", b1a)
}
```
I think it's related to the early return here: https://github.com/holiman/uint256/pull/111/files#diff-af4a4c8b4e176a360fd1472a72691c2ffab2426e94f4e69829a48bd12f01a908R512