@holiman 2022-03-15
Shouldn't it rather be something like this: 
```
	if uLen < dLen {
		copy(rem[:], u[:])
		return rem
	}
```