@chfast 2022-03-17
Should be good now. Do you think we should benchmark this?