@holiman 2022-03-17
> Do you think we should benchmark this? 

I mean, doesn't hurt, but not a blocker for me