@chfast 2022-03-17
It makes it faster (probably by accident).
```
name                      old time/op  new time/op  delta
Div/small/uint256-8       10.7ns ± 0%  10.7ns ± 0%  +0.25%  (p=0.005 n=8+8)
Div/mod64/uint256-8       55.3ns ± 0%  55.0ns ± 0%  -0.50%  (p=0.000 n=9+10)
Div/mod128/uint256-8      81.6ns ± 0%  80.6ns ± 0%  -1.22%  (p=0.000 n=9+9)
Div/mod192/uint256-8      75.2ns ± 0%  74.5ns ± 0%  -0.86%  (p=0.000 n=10+9)
Div/mod256/uint256-8      63.9ns ± 0%  62.6ns ± 0%  -1.93%  (p=0.000 n=9+9)
Mod/small/uint256-8       11.9ns ± 0%  11.9ns ± 0%  -0.22%  (p=0.049 n=8+9)
Mod/mod64/uint256-8       59.0ns ± 0%  58.8ns ± 0%  -0.36%  (p=0.000 n=8+10)
Mod/mod128/uint256-8      85.1ns ± 0%  84.0ns ± 0%  -1.22%  (p=0.000 n=8+9)
Mod/mod192/uint256-8      79.0ns ± 0%  78.5ns ± 0%  -0.63%  (p=0.000 n=8+7)
Mod/mod256/uint256-8      67.9ns ± 0%  67.4ns ± 0%  -0.78%  (p=0.000 n=9+8)
AddMod/small/uint256-8    11.4ns ± 1%  11.4ns ± 0%    ~     (p=0.200 n=10+9)
AddMod/mod64/uint256-8    18.4ns ± 1%  18.1ns ± 0%  -1.67%  (p=0.000 n=10+8)
AddMod/mod128/uint256-8   37.2ns ± 0%  36.9ns ± 0%  -0.88%  (p=0.000 n=9+9)
AddMod/mod192/uint256-8   40.3ns ± 1%  40.0ns ± 0%  -0.71%  (p=0.000 n=10+10)
AddMod/mod256/uint256-8   15.4ns ± 0%  15.4ns ± 0%    ~     (p=0.761 n=8+10)
MulMod/small/uint256-8    28.8ns ± 1%  28.7ns ± 0%    ~     (p=0.502 n=9+8)
MulMod/mod64/uint256-8    67.8ns ± 0%  67.8ns ± 0%    ~     (p=0.481 n=9+10)
MulMod/mod128/uint256-8    103ns ± 0%   102ns ± 0%  -1.15%  (p=0.000 n=8+10)
MulMod/mod192/uint256-8    131ns ± 0%   129ns ± 0%  -1.03%  (p=0.000 n=10+9)
MulMod/mod256/uint256-8    150ns ± 0%   150ns ± 0%    ~     (p=0.135 n=10+10)
MulMod/mod256/uint256r-8  51.5ns ± 0%  51.3ns ± 0%  -0.41%  (p=0.000 n=9+8)
_SDiv/large/uint256-8     83.0ns ± 1%  82.3ns ± 0%  -0.79%  (p=0.000 n=9+9)
[Geo mean]                47.0ns       46.7ns       -0.66%
```